UI update

+++ LionFishWeb/Controllers/UserController.cs
@@ -1,7 +1,9 @@
 ﻿using LionFishWeb.Models;
 using LionFishWeb.Repositories;
+using LionFishWeb.Utility;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Web;
 using System.Web.Mvc;
@@ -16,27 +18,37 @@ namespace LionFishWeb.Controllers
             return View();
         }
 
+        public ActionResult Landing()
+        {
+            if (Convert.ToString(Session["CurrentUser"]) == null || Convert.ToString(Session["CurrentUser"]) == "")
+            {
+                return View("~/Views/Home/LogIn.cshtml");
+            }   
+            else
+                return View();
+        }
+
         [HttpPost][ValidateAntiForgeryToken]
         public ActionResult SubmitCode([Bind(Include = "Code, Email")] ConfirmationCode en)
         {
             IUserRepo ur = new UserRepo();
             IConfirmationCodeRepo cr = new ConfirmationCodeRepo();
             ConfirmationCode cc = cr.GetConfirmationCode(en.Email);
-            if (en.Code.Equals(cc.Code))
+            if (!cc.IsPasswordReset && en.Code.Equals(cc.Code))
             {
                 User u = ur.GetUserByEmail(en.Email);
                 u.IsConfirmed = true;
                 ur.UpdateUser(u);
                 cr.DeleteConfirmationCode(en.Email);
-                return View("~/Views/User/Landing.cshtml", u);
+                return View("~/Views/Home/LogIn.cshtml");
             }
             return View();
         }
 
         [HttpPost]
-        public ActionResult ResendCode([Bind(Include = "Code, Email")] ConfirmationCode cc)
+        public async void ResendCodeAsync([Bind(Include = "Code, Email")] ConfirmationCode cc)
         {
-            return View();
+            await EmailSender.Activate(cc.Email, cc.Code, "Activate your account");
         }
     }
 }
